Update MQ2DPSAdv wrong UI Fix.
Rearranged the order of the checks for child members to exist and commented the entry groups. 
Added a return in the CDPSAdvWnd class creation if CheckUI is true to prevent it from loading anything else about the window (thus preventing the crash). 

Rearranged the order of Setting things in CDPSAdvWnd::LoadSettings() and commented the groups. LTopList->bColumnSizable = true was moved here. 

Put some sanity checks through out to prevent the plugin from doing things if WrongUI was true so that it would load loc, save loc, consider the plugin active. Including in the DPSAdvCmd to account for users attempting to manually "show" the window, open "colors" menu (bShowRaidColors), reload (uses loadloc()), save, listsize, or copy another users location file. If the command is "Show" and WrongUI is true it will output that they have a wrong UI file loaded and suggest some locations to check to remove the old file.