HistoryLimit - Bool should be Int
When saving the HistoryLimit it's currently saving it as a bool. So it's always 1. Meaning every time someone loaded DPSAdv it was always going to be one. 
Appears that it was accidentally changed to a bool 7dec83cbd7f13d53e8bb943546cfcb43f04b7972 here. When fixing merge clobber from legacy to next. Previously it was a string.